Unsnake the project name
While sticking to naming conventions can be helpful, it's not always the most important thing. When it comes to branding and usability. Especially from the command line. It's often better to have a name that's easy to remember and unique. So, going against the Rust naming convention of using snake case will help it stand out and be more memorable. And ultimately easy to remember. If the application was a library/crate first I would be more then happy to stick with the standardization. But due to the reasons I mentioned above and the fact that the project puts more emphasis on the user experience and cli nature. We're going to stick with the existing format.